backend/6001-api-gateway/requirements.txt
fastapi
uvicorn
pydantic
grpcio
grpcio-tools
transformers
torch --index-url https://download.pytorch.org/whl/cpu
